Abstract out podcast formatters
