+++ src/utils/podcastFormatter.ts
@@ -1,0 +1,21 @@
+export function formatDate(isoString: string): string {
+  return new Date(isoString).toLocaleDateString("en-US", {
+    day: "numeric",
+    month: "short",
+    year: "numeric",
+    timeZone: "UTC",
+  });
+}
+
+export function formatDuration(duration: string): string {
+  const parts = duration.split(":");
+  if (parts.length === 3) {
+    const h = parseInt(parts[0], 10);
+    const m = parseInt(parts[1], 10);
+    return h > 0 ? `${h}h ${m}m` : `${m} min`;
+  }
+  if (parts.length === 2) {
+    return `${parseInt(parts[0], 10)} min`;
+  }
+  return duration;
+}
